cli: Split the module
Making a package out of this will make it easier to factorize some code
and introduce per-queue commands.
